Refactor MasterInfo build, support masters command
Moved the code which builds the MasterInfo struct to it's own method in
favor of DRY when adding the SentinelMasters command which will return a
list of Masters being managed.  I'm not sure if it should return a list
or if it should return a map[string]MasterInfo. Leaning toward the
latter.